refactor(frontend): reorganise & rename drought view layers
Split drought view layers into `droughtOptions` and `droughtRegions`.
