+++ frontend/src/config/droughtOptions/styles.ts
