+++ frontend/src/config/droughtRegions/styles.ts
@@ -1,0 +1,8 @@
+import { makeConfig } from 'lib/helpers';
+
+export const DROUGHT_STYLES = makeConfig([
+  {
+    id: 'adaptation',
+    label: 'Adaptation Options',
+  },
+]);
